increasing `messageId` randomness
We've started seeing more and more of the same messageIds coming
through-and the main part of the problem is that Math.random()
isn't a strong source of randomness since it's purely derived
from the timestamp.

This commit makes sure that each individual message will have
stronger randomness (by hashing the message itself) as well as
adding a nonce for multiple messages sent by the same client
in the same ms.

New message ids follow the form:

  "ajs-" + md5(json.stringify(msg) + uuid())

We'll also start prefixing the library name so individual
messages are easy to track down and correlate.