Be less aggressive with retries. (#40)
Before this change, we retry each individual message upto 45 times in case of an error. Since we don't distinguish between error types and retry all errors, In certain cases (particularly when the requests fail due to  adblockers), this retry behaviour is overly aggressive and not desirable.

By reducing the number of retries, we can ensure that we still retry in the average case (i.e. flaky connections) but give up after a reasonable number of events. The downside is that in some cases, legitimate issues that cause us to burn through 10 attempts will not be retried. We're making the tradeoff of having better performance at the cost of slightly worse deliverability.

As a follow up to this, we should also mark certain errors as non-retryable (as per https://paper.dropbox.com/doc/Analytics-library-guidelines-2trBhLKQD4Soz4VatvuGR#:uid=189560039280582553736180&h2=Handling-Network-Errors) so for errors that are indeed intermittent, we can have a more aggresive retry policy.