update send-json with timeout support
Bump the send-json dependency to v4 so we can use the new timeout functionality. Currently we only set a timeout on the retry path to preserve legacy behaviour, though it would be a good idea to set a value here in the future as well.

Note - we need the changes from v4.0.1 (https://github.com/segmentio/send-json/blob/master/HISTORY.md#401--2018-04-16).

Another improvement for the future would be to be able to set per project timeout values so customers can pick and choose. For now, we're using a default of 10 seconds to minimize changes to this code path.