Set timeout on XHR requests when using retry functionality. (#42)
It's a good idea to set a timeout for HTTP requests. Setting a timeout will help avoid resource contention when the network is flaky.

Originally I was going to make this change in send-json, but this requires an update to three repos, https://github.com/segmentio/send-json/pull/2, https://github.com/segment-integrations/analytics.js-integration-segmentio/pull/41 and https://github.com/segmentio/analytics.js-private/pull/227. And although the functionality is behind gated code, it could still be risky.

Instead in this PR, since send-json v3 doesn't support timeouts and it isn't easy to add it without breaking backward compatibility, I've copied the code and tests from send-json and added timeout support. This also makes it easier to run the new code under two specific conditions, 1) the customer has opted into the retry queue functionality and 2) the browser supports CORS.

I think this approach will make it safer to rollout this change, and also easier to rollback in case something goes wrong.

The tests for this change were also conflicting with existing tests. Daniel pointed out that  we were using a fake XMLHttpRequest from sinon, but not cleaning up after tests. As part of this change I updated the existing tests to clean up after themselves.