Enable retryQueue by default. (#49)
Now that we've graduated the feature out of beta, we should enable this by default for all customers.

Existing customers will have this flag set to false (LIB-608), so that they don't notice a change in behaviour after this lands.

To make sure we test both cases well, I moved the existing tests into a table driven test. There is some redundancy with the e2e tests as a result, but I wanted to avoid changing too much logic.