Add flag and logic to delete cross domain identifiers (#54)
* Add flag and logic to delete cross domain identifiers

Some customers may want us to delete cross domain analytics related identifiers when it is disabled. This PR adds a new flag `deleteCrossDomainId` that is checked before running this new logic.

We delete  seg_xid (and metadata) from cookies and  crossDomainId from traits in localStorage. The deletion logic is run only if deletion is enabled for this project, and when either the seg_xid cookie or crossDomainId localStorage trait exists.

To delete the localStorage traits, we use an internal API in analytics.js. This minimizes the risk of introducing a bug that could arise due to interacting with lower level localStorage APIs directly.

Ref: https://segment.atlassian.net/browse/LIB-752

* update method name from deleteCrossDomainId() -> deleteCrossDomainIdIfNeeded()`

* Independently check and delete cookie and localStorage traits

* add more tests