save cross domain identifier cookies from the server as an option (#56)
* remove code to migrate legacy cookies

we've verified that the legacy cookies are not in use anymore, hence we can get rid of this migration logic to simplify the client side code.

here's a link to the same change we made on the server (internal link) https://github.com/segmentio/xid/pull/12.

* save cross domain identifier cookies from the server as an option

Previously xid metadata was stored as client side cookies. This change allows us to set the cookies from a server as httpOnly cookies. We also store the identifier in localStorage To allow the current domain to read it from javascript. This is only set if the request completes succesfully. This behaviour is behind a flag `saveCrossDomainIdInLocalStorage` that is off by default.

This also removes some of the metadata that we don't use (such as the domain of the cookie and timestamp of the cookie)